fix: add axis to context when m=1 (LAS)
When m=1, the sampling process did not enter into the if statement at LOC 151, and the context tensor had only 2 dimensions (lenght of the sequence, dimension of each element in the sequence). We need the third dimension (batch size), which was being added only if m>1 when calling the repeat() function on the context tensor.